Refactoring form template actions and UI to be part of the form template feature folder
