+++ features/form-templates/application/update-template-status.action.ts
@@ -1,23 +1,37 @@
-"use server";
+'use server';
 
-import { ensureAuthenticated } from "@/features/auth";
+import { revalidatePath } from "next/cache";
 import { updateFormTemplate } from "@/services/api";
-import { revalidatePath } from 'next/cache';
+
+interface UpdateTemplateStatusResult {
+  success: boolean;
+  error?: string;
+}
 
 export async function updateTemplateStatusAction(
-  templateId: string,
-  isEnabled: boolean,
-) {
-  await ensureAuthenticated();
-
+  templateId: string, 
+  isEnabled: boolean
+): Promise<UpdateTemplateStatusResult> {
   try {
-    await updateFormTemplate(templateId, { isEnabled });
-    revalidatePath(
-      `/forms/form-templates`,
-    );
+    if (!templateId) {
+      return { success: false, error: "Template ID is required" };
+    }
+
+    // Update the template status
+    await updateFormTemplate(templateId, {
+      isEnabled
+    });
+
+    // Revalidate the template page and templates list
+    revalidatePath(`/forms/templates/${templateId}`);
+    revalidatePath('/forms/templates');
+    
     return { success: true };
   } catch (error) {
-    console.error("Failed to update form template status", error);
-    return { success: false, error: "Failed to update form template status" };
+    console.error("Error updating template status:", error);
+    return { 
+      success: false, 
+      error: `Failed to update template status: ${(error as Error).message}`
+    };
   }
-}
+}
